+++ proxy/grpcproxy/adapter/chan_stream.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package grpcproxy
+package adapter
 
 import (
 	"golang.org/x/net/context"
